% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifs.R
\name{bic_zifs}
\alias{bic_zifs}
\title{BIC for ZIFS Model}
\usage{
bic_zifs(n_0, x, sample_size, p, beta)
}
\arguments{
\item{n_0}{Number of zeros in the sample.}

\item{x}{Vector of non-zero observations.}

\item{sample_size}{Total sample size.}

\item{p}{Parameter (0 < p < 1).}

\item{beta}{Zero-inflation parameter (0 <= beta < 1).}
}
\value{
BIC value.
}
\description{
Computes Bayesian Information Criterion.
}
