% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifs.R
\name{dzifs}
\alias{dzifs}
\title{Zero-Inflated Flory-Schulz Probability Mass Function}
\usage{
dzifs(x, p, beta)
}
\arguments{
\item{x}{Non-negative integer value(s).}

\item{p}{Parameter satisfying 0 < p < 1.}

\item{beta}{Zero-inflation parameter satisfying 0 <= beta < 1.}
}
\value{
Probability mass at x.
}
\description{
Computes the probability mass function (PMF) of the
Zero-Inflated Flory-Schulz (ZIFS) distribution.
}
\examples{
dzifs(1, 0.8, 0.5)
dzifs(0, 0.3, 0.75)

}
