% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifs.R
\name{ll_zifs}
\alias{ll_zifs}
\title{Log-Likelihood of the ZIFS Distribution}
\usage{
ll_zifs(n_0, x, p, beta)
}
\arguments{
\item{n_0}{Number of zeros in the sample.}

\item{x}{Vector of non-zero observations.}

\item{p}{Parameter (0 < p < 1).}

\item{beta}{Zero-inflation parameter (0 <= beta < 1).}
}
\value{
Log-likelihood value.
}
\description{
Computes the log-likelihood for observed data.
}
