% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifs.R
\name{MLE_ZIFS}
\alias{MLE_ZIFS}
\title{Maximum Likelihood Estimation for ZIFS}
\usage{
MLE_ZIFS(input_data)
}
\arguments{
\item{input_data}{Numeric vector of observations.}
}
\value{
Named vector with estimates.
}
\description{
Computes MLE estimators for p and beta.
}
