% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifs.R
\name{modezifs}
\alias{modezifs}
\title{Mode of the ZIFS Distribution}
\usage{
modezifs(p, beta)
}
\arguments{
\item{p}{Parameter (0 < p < 1).}

\item{beta}{Zero-inflation parameter (0 <= beta < 1).}
}
\value{
Mode value. May return two values in bimodal case.
}
\description{
Computes the mode of the ZIFS distribution.
}
\examples{
modezifs(0.4, 0.3)

}
