% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifs.R
\name{mom_ZIFS}
\alias{mom_ZIFS}
\title{Method of Moments Estimation for ZIFS}
\usage{
mom_ZIFS(input_data)
}
\arguments{
\item{input_data}{Numeric vector of observations.}
}
\value{
Named vector with estimates.
}
\description{
Computes method-of-moments estimators for p and beta.
}
