% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifs.R
\name{pzifs}
\alias{pzifs}
\title{Zero-Inflated Flory-Schulz Cumulative Distribution Function}
\usage{
pzifs(x, p, beta)
}
\arguments{
\item{x}{Numeric value.}

\item{p}{Parameter satisfying 0 < p < 1.}

\item{beta}{Zero-inflation parameter satisfying 0 <= beta < 1.}
}
\value{
Cumulative probability P(X <= x).
}
\description{
Computes the cumulative distribution function (CDF)
of the ZIFS distribution.
}
\examples{
pzifs(1, 0.8, 0.5)
pzifs(0.4, 0.3, 0.75)

}
