% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifs.R
\name{qzifs}
\alias{qzifs}
\title{Quantile Function of the ZIFS Distribution}
\usage{
qzifs(r, p, beta)
}
\arguments{
\item{r}{Probability value (0 <= r < 1).}

\item{p}{Parameter (0 < p < 1).}

\item{beta}{Zero-inflation parameter (0 <= beta < 1).}
}
\value{
Quantile corresponding to probability r.
}
\description{
Computes the quantile function of the ZIFS distribution
using the Lambert W function.
}
\examples{
qzifs(0.7, 0.4, 0.3)

}
