% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifs.R
\name{rzifs}
\alias{rzifs}
\title{Random Generation from the ZIFS Distribution}
\usage{
rzifs(k, p, beta)
}
\arguments{
\item{k}{Sample size.}

\item{p}{Parameter (0 < p < 1).}

\item{beta}{Zero-inflation parameter (0 <= beta < 1).}
}
\value{
Numeric vector of random values.
}
\description{
Generates random observations from the ZIFS distribution.
}
