% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifs.R
\name{szifs}
\alias{szifs}
\title{Zero-Inflated Flory-Schulz Survival Function}
\usage{
szifs(x, p, beta)
}
\arguments{
\item{x}{Numeric value.}

\item{p}{Parameter satisfying 0 < p < 1.}

\item{beta}{Zero-inflation parameter satisfying 0 <= beta < 1.}
}
\value{
Survival probability.
}
\description{
Computes the survival function P(X > x) of the ZIFS distribution.
}
\examples{
szifs(1, 0.8, 0.5)

}
