% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifs.R
\name{zifs_iod}
\alias{zifs_iod}
\title{Index of Dispersion of the ZIFS Distribution}
\usage{
zifs_iod(beta, p)
}
\arguments{
\item{beta}{Zero-inflation parameter (0 <= beta < 1).}

\item{p}{Parameter (0 < p < 1).}
}
\value{
Index of dispersion value.
}
\description{
Computes the index of dispersion (variance/mean).
}
\examples{
zifs_iod(0.3, 0.4)

}
