% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifs.R
\name{zifs_mean}
\alias{zifs_mean}
\title{Mean of the ZIFS Distribution}
\usage{
zifs_mean(beta, p)
}
\arguments{
\item{beta}{Zero-inflation parameter (0 <= beta < 1).}

\item{p}{Parameter (0 < p < 1).}
}
\value{
Mean value.
}
\description{
Computes the theoretical mean of the ZIFS distribution.
}
\examples{
zifs_mean(0.3, 0.4)

}
