% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zifs.R
\name{zifs_variance}
\alias{zifs_variance}
\title{Variance of the ZIFS Distribution}
\usage{
zifs_variance(beta, p)
}
\arguments{
\item{beta}{Zero-inflation parameter (0 <= beta < 1).}

\item{p}{Parameter (0 < p < 1).}
}
\value{
Variance value.
}
\description{
Computes the theoretical variance of the ZIFS distribution.
}
\examples{
zifs_variance(0.3, 0.4)

}
